Refactored to use MVVM
